+++ src/main/java/com/crowdin/cli/client/ProjectClient.java
@@ -27,7 +27,11 @@ import java.util.List;
 
 public interface ProjectClient extends Client {
 
-    CrowdinProjectFull downloadFullProject();
+    default CrowdinProjectFull downloadFullProject() {
+        return this.downloadFullProject(null);
+    }
+
+    CrowdinProjectFull downloadFullProject(String branchName);
 
     CrowdinProject downloadProjectWithLanguages();
 
@@ -39,7 +43,7 @@ public interface ProjectClient extends Client {
 
     List<Branch> listBranches();
 
-    Long uploadStorage(String fileName, InputStream content);
+    Long uploadStorage(String fileName, InputStream content) throws ResponseException;
 
     Directory addDirectory(AddDirectoryRequest request) throws ResponseException;