+++ src/test/java/com/crowdin/cli/commands/actions/ListProjectActionTest.java
@@ -15,7 +15,6 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
@@ -45,31 +44,14 @@ public class ListProjectActionTest {
             .minimalBuiltPropertiesBean("*", Utils.PATH_SEPARATOR + "%original_file_name%-CR-%locale%")
             .setBasePath(project.getBasePath());
         pb = pbBuilder.build();
-        when(client.downloadFullProject())
+        when(client.downloadFullProject(null))
             .thenReturn(ProjectBuilder.emptyProject(Long.parseLong(pb.getProjectId()))
                 .addFile("first.po", "gettext", 101L, null, null).build());
 
         action = new ListProjectAction(false, null, true, false);
         action.act(Outputter.getDefault(), pb, client);
 
-        verify(client).downloadFullProject();
-        verifyNoMoreInteractions(client);
-    }
-
-    @Test
-    public void testForNonexistentBranch() throws ResponseException {
-        NewPropertiesWithFilesUtilBuilder pbBuilder = NewPropertiesWithFilesUtilBuilder
-                .minimalBuiltPropertiesBean("*", Utils.PATH_SEPARATOR + "%original_file_name%-CR-%locale%")
-                .setBasePath(project.getBasePath());
-        pb = pbBuilder.build();
-        when(client.downloadFullProject())
-            .thenReturn(ProjectBuilder.emptyProject(Long.parseLong(pb.getProjectId()))
-                .addFile("first.po", "gettext", 101L, null, null).build());
-
-        action = new ListProjectAction(false, "nonexistentBranch", false, false);
-        assertThrows(RuntimeException.class, () -> action.act(Outputter.getDefault(), pb, client));
-
-        verify(client).downloadFullProject();
+        verify(client).downloadFullProject(null);
         verifyNoMoreInteractions(client);
     }
 
@@ -79,7 +61,7 @@ public class ListProjectActionTest {
                 .minimalBuiltPropertiesBean("*", Utils.PATH_SEPARATOR + "%original_file_name%-CR-%locale%")
                 .setBasePath(project.getBasePath());
         pb = pbBuilder.build();
-        when(client.downloadFullProject())
+        when(client.downloadFullProject("existentBranch"))
             .thenReturn(ProjectBuilder.emptyProject(Long.parseLong(pb.getProjectId()))
                 .addFile("first.po", "gettext", 101L, null, null)
                 .addBranches(1L, "existentBranch").build());
@@ -87,7 +69,7 @@ public class ListProjectActionTest {
         action = new ListProjectAction(false, "existentBranch", false, false);
         action.act(Outputter.getDefault(), pb, client);
 
-        verify(client).downloadFullProject();
+        verify(client).downloadFullProject("existentBranch");
         verifyNoMoreInteractions(client);
     }
 }